+++ functions/identityCrossSignVerify.d.ts
@@ -1,0 +1,8 @@
+declare const identityCrossSignVerify: {
+  signCrossSign: (...args: unknown[]) => object;
+  verifyCrossSignObject: (...args: unknown[]) => object;
+  pubkeysMatch: (a: unknown, b: unknown) => boolean;
+  [key: string]: unknown;
+};
+
+export = identityCrossSignVerify;